Onxp 1648 proceed of mint and burn transaction should be implemented

+++ back-end-projects/OntSynHandler/src/main/java/com/github/ontio/model/dao/Oep8.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import javax.persistence.Column;
 import javax.persistence.Id;
 import javax.persistence.Table;
+
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Table(name = "tbl_oep8")
@@ -32,7 +34,7 @@ public class Oep8 {
      * OEP8代币总量
      */
     @Column(name = "total_supply")
-    private Long totalSupply;
+    private BigDecimal totalSupply;
 
     /**
      * OEP8代币符号
@@ -60,7 +62,7 @@ public class Oep8 {
     private Date updateTime;
 
     @Builder
-    public Oep8(String contractHash, String tokenId, String name, Long totalSupply, String symbol, Integer decimals, Date createTime, Boolean auditFlag, Date updateTime) {
+    public Oep8(String contractHash, String tokenId, String name, BigDecimal totalSupply, String symbol, Integer decimals, Date createTime, Boolean auditFlag, Date updateTime) {
         this.contractHash = contractHash;
         this.tokenId = tokenId;
         this.name = name;
@@ -131,7 +133,7 @@ public class Oep8 {
      *
      * @return total_supply - OEP8代币总量
      */
-    public Long getTotalSupply() {
+    public BigDecimal getTotalSupply() {
         return totalSupply;
     }
 
@@ -140,7 +142,7 @@ public class Oep8 {
      *
      * @param totalSupply OEP8代币总量
      */
-    public void setTotalSupply(Long totalSupply) {
+    public void setTotalSupply(BigDecimal totalSupply) {
         this.totalSupply = totalSupply;
     }
